so I am probably doing things the hard way, but I am trying to retain my React learning whilst also trying out a bit of D3 (I want to start making some graphs).
It's been good to play with the useEffect hook and also see the difference between using functions and classes in regards to hooks.

I am still unsure why we now have hooks to provide functions with features that are already in the React.Component class, but it seems to be something people use.

Anyways, I might try and make my graph project with just npm and d3, then come back to this react stuff.

Still have some FCC projects to finish in React, so I don't think I am just going to forget it all. :P