trying out d3 in a component class by following this: https://medium.com/@varvara.munday/d3-in-react-a-step-by-step-tutorial-cba33ce000ce
I am working through https://www.freecodecamp.org/news/how-to-get-started-with-d3-and-react-c7da74a5bd9f/ but I think because it's a couple of years old, syntax/structure may have changed?

Seems to either not work, or generate errors.

It seems that you need to use the useRef hook if you're using a function, or achieve a similar thing in your constructor method if you're using a class.

Once that's sorted, it seems to work as expected (so long as you're referencing the right object).

I am sure a lot of this will make sense once I get into a bit of D3; finding out how that is structured. Looks pretty cool.